linted navbar file
